requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.2
django-datatables-view==1.20.0
django-filter==24.2
sqlparse==0.5.0
